feat: Earning과 Bet 클래스를 추가하여 배팅 및 수익 관리 로직 분리

+++ src/main/java/controller/BlackJackController.java
@@ -91,7 +91,7 @@ public class BlackJackController {
         outputView.displayProfitMessage();
         outputView.displayParticipantProfit("딜러", bettingSession.getDealerProfit());
         for (Player player : players) {
-            int profit = bettingSession.getPlayerProfit(player);
+            double profit = bettingSession.getPlayerProfit(player);
             outputView.displayParticipantProfit(player.getName(), profit);
         }
     }
